function/audio_to_chunk_list/requirements.txt
ffmpeg-python==0.1.17
google-cloud-storage
numpy
futsu
